+++ src/dsnotificationinterface.cpp
@@ -3,7 +3,7 @@
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
 #include "dsnotificationinterface.h"
-#include "darksend.h"
+#include "privatesend-client.h"
 #include "instantx.h"
 #include "governance.h"
 #include "masternodeman.h"
@@ -21,7 +21,7 @@ CDSNotificationInterface::~CDSNotificationInterface()
 void CDSNotificationInterface::UpdatedBlockTip(const CBlockIndex *pindex)
 {
     mnodeman.UpdatedBlockTip(pindex);
-    darkSendPool.UpdatedBlockTip(pindex);
+    privateSendClient.UpdatedBlockTip(pindex);
     instantsend.UpdatedBlockTip(pindex);
     mnpayments.UpdatedBlockTip(pindex);
     governance.UpdatedBlockTip(pindex);
