+++ src/privatesend.h
@@ -2,14 +2,16 @@
 // Distributed under the MIT/X11 software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#ifndef DARKSEND_H
-#define DARKSEND_H
+#ifndef PRIVATESEND_H
+#define PRIVATESEND_H
 
-#include "masternode.h"
-#include "wallet/wallet.h"
+#include "chainparams.h"
+#include "primitives/transaction.h"
+#include "pubkey.h"
+#include "utiltime.h"
+#include "tinyformat.h"
 
-class CDarksendPool;
-class CDarksendBroadcastTx;
+class CPrivateSend;
 
 // timeouts
 static const int PRIVATESEND_AUTO_TIMEOUT_MIN       = 5;
@@ -22,27 +24,11 @@ static const int MIN_PRIVATESEND_PEER_PROTO_VERSION = 70206;
 
 static const CAmount PRIVATESEND_COLLATERAL         = 0.001 * COIN;
 static const CAmount PRIVATESEND_ENTRY_MAX_SIZE     = 9;
-static const int DENOMS_COUNT_MAX                   = 100;
 
-static const int DEFAULT_PRIVATESEND_ROUNDS         = 2;
-static const int DEFAULT_PRIVATESEND_AMOUNT         = 1000;
-static const int DEFAULT_PRIVATESEND_LIQUIDITY      = 0;
-static const bool DEFAULT_PRIVATESEND_MULTISESSION  = false;
-
-// Warn user if mixing in gui or try to create backup if mixing in daemon mode
-// when we have only this many keys left
-static const int PRIVATESEND_KEYS_THRESHOLD_WARNING = 100;
-// Stop mixing completely, it's too dangerous to continue when we have only this many keys left
-static const int PRIVATESEND_KEYS_THRESHOLD_STOP    = 50;
-
-extern int nPrivateSendRounds;
-extern int nPrivateSendAmount;
-extern int nLiquidityProvider;
-extern bool fEnablePrivateSend;
-extern bool fPrivateSendMultiSession;
+class CDarksendBroadcastTx;
 
 // The main object for accessing mixing
-extern CDarksendPool darkSendPool;
+// extern CPrivateSend privateSend;
 
 extern std::map<uint256, CDarksendBroadcastTx> mapDarksendBroadcastTxes;
 extern std::vector<CAmount> vecPrivateSendDenominations;
@@ -100,14 +86,7 @@ public:
         txCollateral(CTransaction())
         {}
 
-    CDarkSendEntry(const std::vector<CTxIn>& vecTxIn, const std::vector<CTxOut>& vecTxOut, const CTransaction& txCollateral) :
-        txCollateral(txCollateral)
-    {
-        BOOST_FOREACH(CTxIn txin, vecTxIn)
-            vecTxDSIn.push_back(txin);
-        BOOST_FOREACH(CTxOut txout, vecTxOut)
-            vecTxDSOut.push_back(txout);
-    }
+    CDarkSendEntry(const std::vector<CTxIn>& vecTxIn, const std::vector<CTxOut>& vecTxOut, const CTransaction& txCollateral);
 
     ADD_SERIALIZE_METHODS;
 
@@ -231,11 +210,10 @@ public:
     bool CheckSignature(const CPubKey& pubKeyMasternode);
 };
 
-/** Used to keep track of current status of mixing pool
- */
-class CDarksendPool
+// static class or base class or helper?
+class CPrivateSend
 {
-private:
+protected:
     // pool responses
     enum PoolMessage {
         ERR_ALREADY_HAVE,
@@ -251,7 +229,7 @@ private:
         ERR_MN_LIST,
         ERR_MODE,
         ERR_NON_STANDARD_PUBKEY,
-        ERR_NOT_A_MN,
+        ERR_NOT_A_MN, // not used
         ERR_QUEUE_FULL,
         ERR_RECENT,
         ERR_SESSION,
@@ -282,57 +260,17 @@ private:
         STATUS_ACCEPTED
     };
 
-    mutable CCriticalSection cs_darksend;
-
     // The current mixing sessions in progress on the network
     std::vector<CDarksendQueue> vecDarksendQueue;
-    // Keep track of the used Masternodes
-    std::vector<CTxIn> vecMasternodesUsed;
 
-    std::vector<CAmount> vecDenominationsSkipped;
-    std::vector<COutPoint> vecOutPointLocked;
-    // Mixing uses collateral transactions to trust parties entering the pool
-    // to behave honestly. If they don't it takes their money.
-    std::vector<CTransaction> vecSessionCollaterals;
     std::vector<CDarkSendEntry> vecEntries; // Masternode/clients entries
 
     PoolState nState; // should be one of the POOL_STATE_XXX values
     int64_t nTimeLastSuccessfulStep; // the time when last successful mixing step was performed, in UTC milliseconds
 
-    int nCachedLastSuccessBlock;
-    int nMinBlockSpacing; //required blocks between mixes
-    const CBlockIndex *pCurrentBlockIndex; // Keep track of current block index
-
     int nSessionID; // 0 if no mixing session is active
 
-    int nEntriesCount;
-    bool fLastEntryAccepted;
-
-    std::string strLastMessage;
-    std::string strAutoDenomResult;
-
-    bool fUnitTest;
-
-    CMutableTransaction txMyCollateral; // client side collateral
     CMutableTransaction finalMutableTransaction; // the finalized transaction ready for signing
-
-    /// Add a clients entry to the pool
-    bool AddEntry(const CDarkSendEntry& entryNew, PoolMessage& nMessageIDRet);
-    /// Add signature to a txin
-    bool AddScriptSig(const CTxIn& txin);
-
-    /// Charge fees to bad actors (Charge clients a fee if they're abusive)
-    void ChargeFees();
-    /// Rarely charge fees to pay miners
-    void ChargeRandomFees();
-
-    /// Check for process
-    void CheckPool();
-
-    void CreateFinalTransaction();
-    void CommitFinalTransaction();
-
-    void CompletedTransaction(PoolMessage nMessageID);
 
     /// Get the denominations for a specific amount of dash.
     int GetDenominationsByAmounts(const std::vector<CAmount>& vecAmount);
@@ -342,92 +280,17 @@ private:
     /// Get the maximum number of transactions for the pool
     int GetMaxPoolTransactions() { return Params().PoolMaxTransactions(); }
 
-    /// Is this nDenom and txCollateral acceptable?
-    bool IsAcceptableDenomAndCollateral(int nDenom, CTransaction txCollateral, PoolMessage &nMessageIDRet);
-    bool CreateNewSession(int nDenom, CTransaction txCollateral, PoolMessage &nMessageIDRet);
-    bool AddUserToExistingSession(int nDenom, CTransaction txCollateral, PoolMessage &nMessageIDRet);
-    /// Do we have enough users to take entries?
-    bool IsSessionReady() { return (int)vecSessionCollaterals.size() >= GetMaxPoolTransactions(); }
-
     /// If the collateral is valid given by a client
     bool IsCollateralValid(const CTransaction& txCollateral);
-    /// Check that all inputs are signed. (Are all inputs signed?)
-    bool IsSignaturesComplete();
-    /// Check to make sure a given input matches an input in the pool and its scriptSig is valid
-    bool IsInputScriptSigValid(const CTxIn& txin);
-    /// Are these outputs compatible with other client in the pool?
-    bool IsOutputsCompatibleWithSessionDenom(const std::vector<CTxDSOut>& vecTxDSOut);
-
-    bool IsDenomSkipped(CAmount nDenomValue) {
-        return std::find(vecDenominationsSkipped.begin(), vecDenominationsSkipped.end(), nDenomValue) != vecDenominationsSkipped.end();
-    }
-
-    /// Create denominations
-    bool CreateDenominated();
-    bool CreateDenominated(const CompactTallyItem& tallyItem, bool fCreateMixingCollaterals);
-
-    /// Split up large inputs or make fee sized inputs
-    bool MakeCollateralAmounts();
-    bool MakeCollateralAmounts(const CompactTallyItem& tallyItem);
-
-    /// As a client, submit part of a future mixing transaction to a Masternode to start the process
-    bool SubmitDenominate();
-    /// step 1: prepare denominated inputs and outputs
-    bool PrepareDenominate(int nMinRounds, int nMaxRounds, std::string& strErrorRet, std::vector<CTxIn>& vecTxInRet, std::vector<CTxOut>& vecTxOutRet);
-    /// step 2: send denominated inputs and outputs prepared in step 1
-    bool SendDenominate(const std::vector<CTxIn>& vecTxIn, const std::vector<CTxOut>& vecTxOut);
-
-    /// Get Masternode updates about the progress of mixing
-    bool CheckPoolStateUpdate(PoolState nStateNew, int nEntriesCountNew, PoolStatusUpdate nStatusUpdate, PoolMessage nMessageID, int nSessionIDNew=0);
-    // Set the 'state' value, with some logging and capturing when the state changed
-    void SetState(PoolState nStateNew);
-
-    /// As a client, check and sign the final transaction
-    bool SignFinalTransaction(const CTransaction& finalTransactionNew, CNode* pnode);
-
-    /// Relay mixing Messages
-    void RelayFinalTransaction(const CTransaction& txFinal);
-    void RelaySignaturesAnon(std::vector<CTxIn>& vin);
-    void RelayInAnon(std::vector<CTxIn>& vin, std::vector<CTxOut>& vout);
-    void RelayIn(const CDarkSendEntry& entry);
-    void PushStatus(CNode* pnode, PoolStatusUpdate nStatusUpdate, PoolMessage nMessageID);
-    void RelayStatus(PoolStatusUpdate nStatusUpdate, PoolMessage nMessageID = MSG_NOERR);
-    void RelayCompletedTransaction(PoolMessage nMessageID);
 
     void SetNull();
 
 public:
-    CMasternode* pSubmittedToMasternode;
     int nSessionDenom; //Users must submit an denom matching this
-    int nCachedNumBlocks; //used for the overview screen
-    bool fCreateAutoBackups; //builtin support for automatic backups
 
-    CDarksendPool() :
-        nCachedLastSuccessBlock(0),
-        nMinBlockSpacing(0),
-        fUnitTest(false),
-        txMyCollateral(CMutableTransaction()),
-        nCachedNumBlocks(std::numeric_limits<int>::max()),
-        fCreateAutoBackups(true) { SetNull(); }
-
-    /** Process a mixing message using the protocol below
-     * \param pfrom
-     * \param strCommand lower case command string; valid values are:
-     *        Command  | Description
-     *        -------- | -----------------
-     *        dsa      | Acceptable
-     *        dsc      | Complete
-     *        dsf      | Final tx
-     *        dsi      | Vector of CTxIn
-     *        dsq      | Queue
-     *        dss      | Signal Final Tx
-     *        dssu     | status update
-     * \param vRecv
-     */
-    void ProcessMessage(CNode* pfrom, std::string& strCommand, CDataStream& vRecv);
+    CPrivateSend() { SetNull(); }
 
     void InitDenominations();
-    void ClearSkippedDenominations() { vecDenominationsSkipped.clear(); }
 
     /// Get the denominations for a list of outputs (returns a bitshifted integer)
     int GetDenominations(const std::vector<CTxOut>& vecTxOut, bool fSingleRandomDenom = false);
@@ -437,31 +300,13 @@ public:
 
     CAmount GetMaxPoolAmount() { return vecPrivateSendDenominations.empty() ? 0 : PRIVATESEND_ENTRY_MAX_SIZE * vecPrivateSendDenominations.front(); }
 
-    void SetMinBlockSpacing(int nMinBlockSpacingIn) { nMinBlockSpacing = nMinBlockSpacingIn; }
-
-    void ResetPool();
-
-    void UnlockCoins();
-
     int GetQueueSize() const { return vecDarksendQueue.size(); }
     int GetState() const { return nState; }
     std::string GetStateString() const;
-    std::string GetStatus();
 
     int GetEntriesCount() const { return vecEntries.size(); }
-
-    /// Passively run mixing in the background according to the configuration in settings
-    bool DoAutomaticDenominating(bool fDryRun=false);
-
-    void CheckTimeout();
-    void CheckForCompleteQueue();
-
-    /// Process a new block
-    void NewBlock();
-
-    void UpdatedBlockTip(const CBlockIndex *pindex);
 };
 
-void ThreadCheckDarkSendPool();
+void ThreadCheckPrivateSend();
 
 #endif
