+++ src/qt/coincontroldialog.cpp
@@ -19,8 +19,8 @@
 #include "main.h" // For minRelayTxFee
 #include "wallet/wallet.h"
 
-#include "darksend.h"
 #include "instantx.h"
+#include "privatesend-client.h"
 
 #include <boost/assign/list_of.hpp> // for 'map_list_of()'
