+++ src/qt/transactionrecord.cpp
@@ -11,8 +11,8 @@
 #include "timedata.h"
 #include "wallet/wallet.h"
 
-#include "darksend.h"
 #include "instantx.h"
+#include "privatesend.h"
 
 #include <stdint.h>
