+++ src/qt/walletmodel.cpp
@@ -20,9 +20,9 @@
 #include "wallet/wallet.h"
 #include "wallet/walletdb.h" // for BackupWallet
 
-#include "darksend.h"
 #include "instantx.h"
 #include "spork.h"
+#include "privatesend-client.h"
 
 #include <stdint.h>
